==> analysis/bus-lane-analysis/scratch_notebooks/03_create_bus_lane_segments_by_intersection.ipynb <==
# copied from a previous rawnav repo - needs to be updated for the new workflow

import os, sys, glob, pandas as pd, geopandas as gpd
import pyarrow as pa, pyarrow.parquet as pq
import folium
from folium.plugins import MarkerCluster
from folium import CircleMarker
from IPython.display import IFrame

#### ben imports
import json
import numpy as np
from scipy.spatial import cKDTree
from shapely.geometry import LineString, Point


pd.set_option('display.max_columns', None)
if not sys.warnoptions:
    import warnings
    warnings.simplefilter("ignore")
%load_ext autoreload
%autoreload 2

#NAD83 Maryland (ft)
wmata_crs = 2248

from ipywidgets import interact
from shapely.ops import cascaded_union, linemerge

#os.path.pardir is equivalent to ".."
#need to append folder 3 levels up in order to load wmatarawnav
sys.path.append(os.path.join(os.path.pardir, os.path.join( os.path.pardir,os.path.pardir)))
import wmatarawnav as wr

# 1.2 Set Global Parameters
############################################
if os.getlogin() == "E048374":
    # Working Paths
    path_working = r"C:\Users\E048374\OneDrive - WMATA\rawnav_rachel_fork\WMATA_AVL"
    os.chdir(os.path.join(path_working))
    sys.path.append(r"C:\Users\E048374\OneDrive - WMATA\rawnav_rachel_fork\WMATA_AVL")
    path_source_data = r"\\l-600730\RawNavArchive"
    path_sp = r"C:\Users\E048374\Documents\RawNav"
    path_processed_data = os.path.join(path_sp, "data", "02-processed")
    path_segments = path_processed_data
    path_sched_data = os.path.join(path_working, "data", "02-processed")
    
else:
    raise FileNotFoundError("Define the path_working, path_source_data, gtfs_dir, \
                            ZippedFilesloc, and path_processed_data in a new elif block")

## load segments for visualization

segments_loc = "C:\\Users\\E048374\\OneDrive - WMATA\\rawnav_rachel_fork\\WMATA_AVL\\data\\02-processed\\2020-09-08-segments_14th_only.geojson"
segments = gpd.read_file(segments_loc)

## load bus stops on south west side of 14th for southbound route

pattern_id = '5201'
bus_sched_df = (
    pd.read_csv(
        os.path.join(path_sched_data, f"bus_sched_{pattern_id}.csv")
        ,dtype={'pattern':'int32','route':str,'pattern_id':str}
    )
)

stop_seq = gpd.GeoDataFrame(
                    bus_sched_df, 
                    geometry = gpd.points_from_xy(
                        bus_sched_df.stop_lon,
                        bus_sched_df.stop_lat
                    ),
    crs='EPSG:4326'
)

stop_seq.dtypes

#need to change datatype for merge with stop sequence below
stop_seq.loc[:,'pattern_id'].unique()

stop_seq_subset = stop_seq[stop_seq.loc[:,'pattern_id']=='5201'].copy()

def get_centroid_gpd(gpd_df):
    '''
    '''
    minx = gpd_df.bounds.mean()['minx']
    miny = gpd_df.bounds.mean()['miny']
    maxx = gpd_df.bounds.mean()['maxx']
    maxy = gpd_df.bounds.mean()['maxy']
    avg_x = np.array([minx, maxx]).mean()
    avg_y = np.array([miny, maxy]).mean()
    #return Lat, Long
    return[avg_y, avg_x]

esri_imagery = "https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}"
esri_attribution = \
    "Tiles &copy; Esri &mdash; Source: Esri, i-cubed, USDA, USGS, AEX, GeoEye, Getmapping, Aerogrid, IGN, IGP, " \
    "UPR-EGP, and the GIS User Community"
this_map = folium.Map(location=get_centroid_gpd(stop_seq_subset),
                       tiles='cartodbdark_matter', zoom_start=16, max_zoom=25, control_scale=True)
folium.TileLayer(name="EsriImagery", tiles=esri_imagery, attr=esri_attribution,
                 zoom_start=16, max_zoom=25, control_scale=True).add_to(this_map)
folium.TileLayer('cartodbpositron', zoom_start=16, max_zoom=20, control_scale=True).add_to(this_map)
folium.TileLayer('openstreetmap', zoom_start=16, max_zoom=20, control_scale=True).add_to(this_map)

folium.Choropleth(
    segments,
    line_weight=3,
    line_color='blue'
).add_to(this_map)

h = folium.FeatureGroup(name='stop locations')

wr.plot_marker_clusters(this_map, stop_seq_subset, "stop_lat","stop_lon",h, fill_color='#f79e19')

this_map.add_children(h)

# select closest stops to bus lane
there's a better, programmatic, way to do this but I just looked at the map, clicked on the close stops and I'm filtering for them below :)

geoid_list = [16203, 16662, 7243]
stop_seq_bus_lane = stop_seq_subset[stop_seq_subset.loc[:,'geoid'].isin(geoid_list)].copy()

geoid_segment_dict = {16203:'14th_12', 16662:'14th_14', 7243:'14th_25'}

segment_geoid_dict = dict()
for keys, values in geoid_segment_dict.items():
    segment_geoid_dict[values] = keys

stop_seq_bus_lane.loc[:,'seg_name_id'] = stop_seq_bus_lane.loc[:,'geoid'].apply(lambda x: geoid_segment_dict[x])

segments_w_stops = segments.merge(stop_seq_bus_lane[['seg_name_id','geoid','stop_id','route','pattern']]
               , how='left'
              ,on='seg_name_id').copy()

segments_w_stops.loc[:,'geoid'] = segments_w_stops.loc[:,'geoid'].apply(lambda x: int(x) if x==x else 0).copy()
segments_w_stops.loc[:,'stop_id'] = segments_w_stops.loc[:,'stop_id'].apply(lambda x: int(x) if x==x else 0).copy()
segments_w_stops.loc[:,'route'] = '52'
segments_w_stops.loc[:,'pattern'] = 1
segments_w_stops

segments_w_stops.to_crs(2248, inplace=True)
segments_w_stops.loc[:,'length'] = segments_w_stops.loc[:,'geometry'].apply(lambda x: x.length)

# segments_test = segments_w_stops[segments_w_stops.loc[:,'seg_name_id'].isin(['14th_22','14th_24','14th_25'])].copy()
# segments_test.to_crs(2248, inplace=True)
# segments_test.loc[:,'length'] = segments_test.loc[:,'geometry'].apply(lambda x: x.length)

segments_w_stops.to_crs("EPSG:4326", inplace=True)
segments_w_stops.reset_index(inplace=True, drop=True)

file_name = "seg_5201_by_intersection.geojson"
file_loc = f"C:\\Users\\E048374\\OneDrive - WMATA\\rawnav_rachel_fork\\WMATA_AVL\\data\\02-processed\\{file_name}"
segments_w_stops.to_file(file_loc, driver='GeoJSON')

segments_w_stops

stop_seq_pattern_seg = segments_w_stops[['seg_name_id','geoid','stop_id','route','pattern']].copy()
stop_seq_pattern_seg

file_name = "stop_seq_pattern_5201_by_intersection.csv"
file_loc = f"C:\\Users\\E048374\\OneDrive - WMATA\\rawnav_rachel_fork\\WMATA_AVL\\data\\02-processed\\{file_name}"
stop_seq_pattern_seg.to_csv(file_loc, index=False)

## Old consolidation method

consolidated_segments_df = segments_test[segments_test.loc[:,'seg_name_id']=='14th_25'][['seg_name_id'
                                                                                         ,'name_str'
                                                                                        ,'geoid', 'stop_id']].copy()

segments_test

consolidated_len = segments_test.loc[:,'length'].sum()
consolidated_geom = linemerge(cascaded_union(segments_test.geometry.values))

consolidated_seg_gdf = gpd.GeoDataFrame(consolidated_segments_df, crs="EPSG:4326", geometry=[consolidated_geom])
consolidated_seg_gdf.loc[:,'length'] = consolidated_len

consolidated_seg_gdf.reset_index(inplace=True, drop=True)

file_name = "seg_5201_by_stop.geojson"
file_loc = f"C:\\Users\\E048374\\OneDrive - WMATA\\rawnav_rachel_fork\\WMATA_AVL\\data\\02-processed\\{file_name}"
consolidated_seg_gdf.to_file(file_loc, driver='GeoJSON')

stop_seq_pattern = stop_seq_pattern_seg[stop_seq_pattern_seg.loc[:,'geoid']==7243].copy()

stop_seq_pattern.reset_index(inplace=True, drop=True)

file_name = "stop_seq_pattern_5201_by_stop.csv"
file_loc = f"C:\\Users\\E048374\\OneDrive - WMATA\\rawnav_rachel_fork\\WMATA_AVL\\data\\02-processed\\{file_name}"
stop_seq_pattern.to_csv(file_loc, index=False)

# Join segment id table with schedule to get geodescription for each stop along the segment
